Re-use Canvas component in TileMapCanvas
