+++ src/components/canvas.css
@@ -1,20 +1,14 @@
 :root {
-  --pixelSize: 24px;
-}
-
-.container {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
+  --pixelSize: 4px;
 }
 
 .canvas {
   border-collapse: collapse;
+  table-layout: fixed;
 }
 
 .canvas td {
-  border: solid 1px #ddd;
+  /* border: solid 1px #ddd; */
   cursor: crosshair;
   height: var(--pixelSize);
   text-align: center;
